+++ ItemUnlocker/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Removed the upgraded versions of starter items from the pool. Can still purchase unupgraded versions.
 
+Fixed Essentials Compatibility
+
 # 1.0.0
 
 Initial release.